CRUDServices: - Performed logging in services + added checks - Added NoContentException where needed - Fixed bug in updating method(when trying to update data with null gender) - Modified tests

+++ src/main/java/com/example/adventureprogearjava/services/impl/ProductAttributeServiceImpl.java
@@ -1,9 +1,8 @@
 package com.example.adventureprogearjava.services.impl;
 
 import com.example.adventureprogearjava.dto.ProductAttributeDTO;
-import com.example.adventureprogearjava.dto.ProductDTO;
-import com.example.adventureprogearjava.entity.Product;
 import com.example.adventureprogearjava.entity.ProductAttribute;
+import com.example.adventureprogearjava.exceptions.NoContentException;
 import com.example.adventureprogearjava.exceptions.ResourceNotFoundException;
 import com.example.adventureprogearjava.mapper.ProductAttributeMapper;
 import com.example.adventureprogearjava.repositories.ProductAttributeRepository;
@@ -11,6 +10,7 @@ import com.example.adventureprogearjava.services.CRUDService;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,6 +19,7 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class ProductAttributeServiceImpl implements CRUDService<ProductAttributeDTO> {
     ProductAttributeRepository productAttributeRepo;
@@ -27,6 +28,7 @@ public class ProductAttributeServiceImpl implements CRUDService<ProductAttribute
 
     @Override
     public List<ProductAttributeDTO> getAll() {
+        log.info("Getting all productAttributes");
         return productAttributeRepo.findAll()
                 .stream()
                 .map(productAttributeMapper::toDto)
@@ -35,6 +37,7 @@ public class ProductAttributeServiceImpl implements CRUDService<ProductAttribute
 
     @Override
     public ProductAttributeDTO getById(Long id) {
+        log.info("Getting productAttributes by id: {}", id);
         Optional<ProductAttribute> productAttribute = productAttributeRepo.findById(id);
         if (productAttribute.isEmpty()) {
             throw new ResourceNotFoundException("Resource is not available!");
@@ -44,6 +47,7 @@ public class ProductAttributeServiceImpl implements CRUDService<ProductAttribute
 
     @Override
     public ProductAttributeDTO create(ProductAttributeDTO productAttributeDTO) {
+        log.info("Creating new productAttribute.");
         productAttributeRepo.save(productAttributeMapper
                 .toEntity(productAttributeDTO));
         return productAttributeDTO;
@@ -52,9 +56,10 @@ public class ProductAttributeServiceImpl implements CRUDService<ProductAttribute
     @Override
     @Transactional
     public void update(ProductAttributeDTO productAttributeDTO, Long id) {
+        log.info("Updating productAttribute with id: {}", id);
         if (!productAttributeRepo.existsById(id)) {
-            productAttributeRepo.save(productAttributeMapper
-                    .toEntity(productAttributeDTO));
+            log.warn("ProductAttribute not found!");
+            throw new ResourceNotFoundException("Resource is not available!");
         } else {
             productAttributeRepo.update(id,
                     productAttributeDTO.getSize(),
@@ -66,6 +71,11 @@ public class ProductAttributeServiceImpl implements CRUDService<ProductAttribute
 
     @Override
     public void delete(Long id) {
+        log.info("Deleting productAttribute with id: {}", id);
+        if (!productAttributeRepo.existsById(id)){
+            log.warn("No content present!");
+            throw new NoContentException("No content present!");
+        }
         productAttributeRepo.deleteById(id);
     }
 }
